Change deliverableartifact#brewbuildid to BIGINT in DB script

+++ dto/src/main/java/org/jboss/pnc/dto/response/AnalyzedArtifact.java
@@ -38,7 +38,7 @@ public class AnalyzedArtifact {
     /**
      * The ID of the Brew build (in case the artifact was built in the Brew) in which was built this artifact.
      */
-    Integer brewId;
+    Long brewId;
 
     /**
      * Artifact's actual data.
